requirements.txt
PyQt5
boto3
python-dotenv
